# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aqarly API",
    "description": "Backend for the Aqarly platform's apps.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request Code",
        "description": "Sends a sign-in code for this phone and app. It's issued even for a\nnumber with no account, so this can't be used to find out who has one.",
        "operationId": "request_code_auth_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify",
        "description": "Proves the code and opens a session. In the field app and the admin\nportals the number must already have an account; in the tenant portal a\nnew number gets a session it can register with.",
        "operationId": "verify_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "A tenant says who they are and which unit they live in. They wait for\ntheir building (ops) to confirm it before they see anything of the unit's.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "trade": {
            "type": "string",
            "enum": [
              "maintenance",
              "housekeeping"
            ],
            "title": "Trade"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "trade"
        ],
        "title": "AdminOut"
      },
...
      "CodeIn": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "app": {
            "type": "string",
            "enum": [
              "tenant",
              "field",
              "ops",
              "housekeeping"
            ],
            "title": "App"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "app"
        ],
        "title": "CodeIn"
      },
      "CodeOut": {
        "properties": {
          "sent": {
            "type": "boolean",
            "title": "Sent"
          },
          "shownCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Showncode"
          }
        },
        "type": "object",
        "required": [
          "sent",
          "shownCode"
        ],
        "title": "CodeOut"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "description": "Every refusal: a message the frontend can show as it is."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeOut": {
        "properties": {
          "app": {
            "type": "string",
            "enum": [
              "tenant",
              "field",
              "ops",
              "housekeeping"
            ],
            "title": "App"
          },
          "kind": {
            "type": "string",
            "enum": [
              "tenant",
              "registration",
              "new",
              "staff",
              "admin"
            ],
            "title": "Kind"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "tenant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantAccountOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "registration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegistrationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "staff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StaffOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "admin": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "app",
          "kind",
          "phone"
        ],
        "title": "MeOut",
        "description": "Who is signed in. `kind` says which of the records below is set:\n`tenant`; `registration` (waiting for ops, or declined); `new` (a proved\nphone with no account yet, which can register); `staff`; `admin`."
      },
...
      "PropertyOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address"
        ],
        "title": "PropertyOut"
      },
...
      "RegisterIn": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "unitId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unitid"
          }
        },
        "type": "object",
        "title": "RegisterIn"
      },
      "RegistrationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "unit": {
            "$ref": "#/components/schemas/UnitOut"
          },
          "property": {
            "$ref": "#/components/schemas/PropertyOut"
          },
          "currentTenant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "approved",
                  "declined"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "decidedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decidedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "unit",
          "property",
          "currentTenant",
          "createdAt",
          "decision",
          "decidedAt"
        ],
        "title": "RegistrationOut"
      },
...
      "SessionOut": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "me": {
            "$ref": "#/components/schemas/MeOut"
          }
        },
        "type": "object",
        "required": [
          "token",
          "me"
        ],
        "title": "SessionOut",
        "description": "A new session. Keep `token` (in an HttpOnly cookie) and send it as the\n`X-Session` header on every call; it isn't shown again."
      },
...
      "StaffOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "role": {
            "type": "string",
            "enum": [
              "maintenance",
              "housekeeping"
            ],
            "title": "Role"
          },
          "photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo"
          },
          "retiredAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retiredat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "role",
          "photo",
          "retiredAt"
        ],
        "title": "StaffOut"
      },
...
      "TenantAccountOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UnitOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "property": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropertyOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "email",
          "unit",
          "property"
        ],
        "title": "TenantAccountOut",
        "description": "A tenant with the unit they occupy."
      },
      "TenantOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "email"
        ],
        "title": "TenantOut"
      },
...
      "UnitOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "propertyId": {
            "type": "string",
            "title": "Propertyid"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "enum": [
              "occupied",
              "under-maintenance",
              "vacant"
            ],
            "title": "Status"
          },
          "tenantId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenantid"
          },
          "bedrooms": {
            "type": "integer",
            "title": "Bedrooms"
          },
          "bathrooms": {
            "type": "integer",
            "title": "Bathrooms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "propertyId",
          "label",
          "status",
          "tenantId",
          "bedrooms",
          "bathrooms"
        ],
        "title": "UnitOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyIn": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "app": {
            "type": "string",
            "enum": [
              "tenant",
              "field",
              "ops",
              "housekeeping"
            ],
            "title": "App"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "app",
          "code"
        ],
        "title": "VerifyIn"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "The token from POST /auth/verify. In Swagger: Authorize, then paste it.",
        "in": "header",
        "name": "X-Session"
      }
    }
  }
}